add event handlers and data models for demo app
