+++ crates/spec-ai-tui/src/style/mod.rs
@@ -4,8 +4,10 @@ mod color;
 mod modifier;
 mod style;
 mod styled;
+pub mod text_utils;
 
 pub use color::Color;
 pub use modifier::Modifier;
 pub use style::Style;
-pub use styled::{Span, Line, Text};
+pub use styled::{Line, Span, Text};
+pub use text_utils::{truncate, wrap_text};
